% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motifs.R
\name{computeMotifs}
\alias{computeMotifs}
\title{Counting the Number of Motifs in RNA or Protein Sequences}
\usage{
computeMotifs(
  seqs,
  seqType = c("RNA", "Pro"),
  motifRNA = c("rpiCOOL", "Fox1", "Nova", "Slm2", "Fusip1", "PTB", "ARE", "hnRNPA1",
    "PUM", "U1A", "HuD", "QKI", "U2B", "SF1", "HuR", "YB1", "AU", "UG", "selected5"),
  motifPro = c("rpiCOOL", "E", "H", "K", "R", "H_R", "EE", "KK", "HR_RH", "RS_SR",
    "RGG", "YGG"),
  newMotif = NULL,
  newMotifOnly = FALSE,
  parallel.cores = 2,
  cl = NULL
)
}
\arguments{
\item{seqs}{sequences loaded by function \code{\link[seqinr]{read.fasta}} from \code{\link[seqinr]{seqinr-package}}. Or a list of RNA/protein sequences.
RNA sequences will be converted into lower case letters, but
protein sequences will be converted into upper case letters.
Each sequence should be a vector of single characters.}

\item{seqType}{a string that specifies the nature of the sequence: \code{"RNA"} or \code{"Pro"} (protein).
If the input is DNA sequence and \code{seqType = "RNA"}, the DNA sequence will be converted to RNA sequence automatically.
Default: \code{"RNA"}.}

\item{motifRNA}{strings specifying the motifs that are counted in RNA sequences. Ignored if \code{seqType = "Pro"}.
Options: \code{"rpiCOOL"}, \code{"selected5"},
\code{"Fox1"}, \code{"Nova"}, \code{"Slm2"}, \code{"Fusip1"}, \code{"PTB"}, \code{"ARE"}, \code{"hnRNPA1"},
\code{"PUM"}, \code{"U1A"}, \code{"HuD"}, \code{"QKI"}, \code{"U2B"}, \code{"SF1"}, \code{"HuR"}, \code{"YB1"},
\code{"AU"}, and \code{"UG"}. Multiple elements can be selected at the same time.
If \code{"rpiCOOL"}, all default motifs will be counted.
\code{"selected5"} indicates the total number of the occurrences of: PUM, Fox-1, U1A, Nova, and ARE which
are regarded as the five most over-presented binding motifs. See details below.}

\item{motifPro}{strings specifying the motifs that are counted in protein sequences. Ignored if \code{seqType = "RNA"}.
Options: \code{"rpiCOOL"}, \code{"E"}, \code{"H"}, \code{"K"}, \code{"R"}, \code{"H_R"},
\code{"EE"}, \code{"KK"}, \code{"HR_RH"}, \code{"RS_SR"},
\code{"RGG"}, and \code{"YGG"}. Multiple elements can be selected at the same time.
\code{"H_R"} indicates the total number of the occurrences of: H and R.
\code{"HR_RH"} indicates the total number of the occurrences of: HR and RH.
\code{"RS_SR"} indicates the total number of the occurrences of: RS and SR.
If \code{"rpiCOOL"}, the default motifs of rpiCOOL (\code{"E"}, \code{"K"}, \code{"H_R"},
\code{"EE"}, \code{"KK"}, \code{"RS_SR"}, \code{"RGG"}, and \code{"YGG"}) will be counted.
See details below.}

\item{newMotif}{list defining new motifs not listed above. New motifs are counted in RNA or protein sequences.
For example, \code{newMotif = list(hnRNPA1 = c("UAGGGU", "UAGGGA"), SF1 = "UACUAAC")}.
This parameter can be used together with parameter \code{motifRNA} or \code{motifPro}. Default: \code{NULL}.}

\item{newMotifOnly}{logical. If \code{TRUE}, only the new motifs defined in \code{newMotif} will be counted.
Default: \code{FALSE}.}

\item{parallel.cores}{an integer specifying the number of cores for parallel computation. Default: \code{2}.
Set \code{parallel.cores = -1} to run with all the cores. \code{parallel.cores} should be == -1 or >= 1.}

\item{cl}{parallel cores to be passed to this function.}
}
\value{
This function returns a data frame. Row names are sequences names, and column names are motif names.
}
\description{
Counts the number of motifs occurring in RNA/protein sequences. Motifs employed by tool "rpiCOOL"
can be selected. New motifs can also be defined.
}
\details{
This function can count the motifs in RNA or protein sequences.

The default motifs are selected or derived from tool "rpiCOOL" (Ref: [2]).

\itemize{
\item Motifs of RNA

\enumerate{
\item Fox1: UGCAUGU;
\item Nova: UCAUUUCAC, UCAUUUCAU, CCAUUUCAC, CCAUUUCAU;
\item Slm2: UAAAC, UAAAA, UAAUC, UAAUA;
\item Fusip1: AAAGA, AAAGG, AGAGA, AGAGG, CAAGA, CAAGG, CGAGA, CGAGG;
\item PTB: UUUUU, UUUCU, UCUUU, UCUCU;
\item ARE: UAUUUAUU;
\item hnRNPA1: UAGGGU, UAGGGA;
\item PUM: UGUAAAUA, UGUAGAUA, UGUAUAUA, UGUACAUA;
\item U1A: AUUGCAC;
\item HuD: UUAUUU;
\item QKI: AUUAAU, AUUAAC, ACUAAU, ACUAAC;
\item U2B: AUUGCAG;
\item SF1: UACUAAC;
\item HuR: UUUAUUU, UUUGUUU, UUUCUUU, UUUUUUU;
\item YB1: CCUGCG, UCUGCG;
\item AU: AU;
\item UG: UG.

If \code{"rpiCOOL"}, all default motifs will be counted, and there is no need to input other default motifs.
\code{"selected5"} indicates the total number of the occurrences of: PUM, Fox-1, U1A, Nova, and ARE which
are regarded as the five most over-represented binding motifs.}

\item Motifs of protein

\enumerate{
\item E: E;
\item H: H;
\item K: K;
\item R: R;
\item EE: EE;
\item KK: KK;
\item HR (\code{"H_R"}): H, R;
\item HR (\code{"HR_RH"}): HR, RH;
\item RS (\code{"RS_SR"}): RS, SR;
\item RGG: RGG;
\item YGG: YGG.

If \code{"rpiCOOL"}, default motifs of rpiCOOL (\code{"E"}, \code{"K"}, \code{"H_R"},
\code{"EE"}, \code{"KK"}, \code{"RS_SR"}, \code{"RGG"}, and \code{"YGG"}) will be counted.
} }

There are some minor differences between this function and the extraction scheme of rpiCOOL.
In this function, motifs will be scanned directly.
As to the extraction scheme of rpiCOOL, some motifs (\code{"UG"}, \code{"AU"}, and \code{"H_R"})
are scanned in a 10 nt/aa sliding-window.

New motif patterns are also supported. Users can pass new patterns to argument "newMotif" as a list.
Format:

\code{newMotif = list(*motif_name* = c("*motif_pattern_1*", "*motif_pattern_2*"))}.

For example: \code{newMotif = list(HR_RH = c("HR", "RH"), RGG = "RGG")}.
"HR_RH" is the name of this motif which contains two patterns: "HR" and "RH".
}
\section{References}{

[1] Han S, Liang Y, Ma Q, \emph{et al}.
ncProR: an integrated R package for effective ncRNA-protein interaction prediction.
(\emph{Submitted})

[2] Akbaripour-Elahabad M, Zahiri J, Rafeh R, \emph{et al}.
rpiCOOL: A tool for In Silico RNA-protein interaction detection using random forest.
J. Theor. Biol. 2016; 402:1-8

[3] Pancaldi V, Bahler J.
In silico characterization and prediction of global protein-mRNA interactions in yeast.
Nucleic Acids Res. 2011; 39:5826-36

[4] Castello A, Fischer B, Eichelbaum K, \emph{et al}.
Insights into RNA Biology from an Atlas of Mammalian mRNA-Binding Proteins.
Cell 2012; 149:1393-1406

[5] Ray D, Kazan H, Cook KB, \emph{et al}.
A compendium of RNA-binding motifs for decoding gene regulation.
Nature 2013; 499:172-177

[6] Jiang P, Singh M, Coller HA.
Computational assessment of the cooperativity between RNA binding proteins and MicroRNAs in Transcript Decay.
PLoS Comput. Biol. 2013; 9:e1003075
}

\examples{
data(demoPositiveSeq)
seqsRNA <- demoPositiveSeq$RNA.positive
seqsPro <- demoPositiveSeq$Pro.positive

motifRNA1 <- computeMotifs(seqsRNA, seqType = "RNA", motifRNA = "rpiCOOL",
                           parallel.cores = 2)

motifRNA2 <- computeMotifs(seqsRNA, seqType = "RNA",
                           motifRNA = c("Fox1", "HuR", "ARE"), parallel.cores = 2)

motifPro1 <- computeMotifs(seqsPro, seqType = "Pro",
                           motifPro = c("rpiCOOL", "HR_RH"), parallel.cores = 2)

# Customized motifs are also supported and can be extracted with default motifs.
# Pass new motif patterns to "newMotif" argument as a list:

motifPro2 <- computeMotifs(seqsPro, seqType = "Pro", motifPro = c("E", "K", "KK"),
                           newMotif = list(HR_RH = c("HR", "RH"), RGG = "RGG"),
                           parallel.cores = 2)

# set "newMotifOnly = TRUE", if compute customized motifs only:

motifPro3 <- computeMotifs(seqsPro, seqType = "Pro",
                           newMotif = list(HR_RH = c("HR", "RH"), RGG = "RGG"),
                           newMotifOnly = TRUE, parallel.cores = 2)
}
\seealso{
\code{\link{featureMotifs}}
}
