% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicochemicalProperty.R
\name{computePhysChem}
\alias{computePhysChem}
\title{Computation of the Physicochemical Features of RNA or Protein Sequences}
\usage{
computePhysChem(
  seqs,
  seqType = c("RNA", "Pro"),
  Fourier.len = 10,
  physchemRNA = c("hydrogenBonding", "vanderWaal"),
  physchemPro = c("polarity.Grantham", "polarity.Zimmerman", "bulkiness.Zimmerman",
    "isoelectricPoint.Zimmerman", "hphob.BullBreese", "hphob.KyteDoolittle",
    "hphob.Eisenberg", "hphob.HoppWoods"),
  as.list = TRUE,
  parallel.cores = 2,
  cl = NULL
)
}
\arguments{
\item{seqs}{sequences loaded by function \code{\link[seqinr]{read.fasta}} from \code{\link[seqinr]{seqinr-package}}. Or a list of RNA/protein sequences.
RNA sequences will be converted into lower case letters, but
protein sequences will be converted into upper case letters.
Each sequence should be a vector of single characters.}

\item{seqType}{a string that specifies the nature of the sequence: \code{"RNA"} or \code{"Pro"} (protein).
If the input is DNA sequence and \code{seqType = "RNA"}, the DNA sequence will be converted to RNA sequence automatically.
Default: \code{"RNA"}.}

\item{Fourier.len}{positive integer specifying the Fourier series length that will be used as features.
The \code{Fourier.len} should be >= the length of the input sequence. Default: \code{10}.}

\item{physchemRNA}{strings specifying the physicochemical properties that are computed in RNA sequences. Ignored if \code{seqType = "Pro"}.
Options: \code{"hydrogenBonding"} for Hydrogen-bonding and \code{"vanderWaal"} for Van der Waal's interaction
Multiple elements can be selected at the same time. (Ref: [2])}

\item{physchemPro}{strings specifying the physicochemical properties that are computed in protein sequences. Ignored if \code{seqType = "RNA"}.
Options: \code{"polarity.Grantham"}, \code{"polarity.Zimmerman"}, \code{"bulkiness.Zimmerman"}, \code{"isoelectricPoint.Zimmerman"},
\code{"hphob.BullBreese"}, \code{"hphob.KyteDoolittle"}, \code{"hphob.Eisenberg"}, and
\code{"hphob.HoppWoods"}.
Multiple elements can be selected at the same time. See details below. (Ref: [3-9])}

\item{as.list}{logical. The result will be returned as a list or a data frame.}

\item{parallel.cores}{an integer that indicates the number of cores for parallel computation. Default: \code{2}.
Set \code{parallel.cores = -1} to run with all the cores. \code{parallel.cores} should be == -1 or >= 1.}

\item{cl}{parallel cores to be passed to this function.}
}
\value{
This function returns a data frame if \code{as.list = FALSE} or returns a list if \code{as.list = TRUE}.
}
\description{
The function \code{computePhysChem} computes the physicochemical features of RNA
or protein sequences.
}
\details{
The default physicochemical properties are selected or derived from tool "catRAPID" (Ref: [10])
and "lncPro" (Ref: [11]).
In "catRAPID", \code{Fourier.len = 50}; in "lncPro", \code{Fourier.len} is set as \code{10}.

\itemize{
\item The physicochemical properties of RNA

\enumerate{
\item Hydrogen-bonding (\code{"hydrogenBonding"}) (Ref: [2])
\item Van der Waal's interaction (\code{"vanderWaal"}) (Ref: [2])
}

\item The physicochemical properties of protein sequence

\enumerate{
\item polarity \code{"polarity.Grantham"} (Ref: [3])
\item polarity \code{"polarity.Zimmerman"} (Ref: [4])
\item bulkiness \code{"bulkiness.Zimmerman"} Ref: [4]
\item isoelectric point \code{"isoelectricPoint.Zimmerman"} (Ref: [4])
\item hydropathicity \code{"hphob.BullBreese"} (Ref: [5])
\item hydropathicity \code{"hphob.KyteDoolittle"} (Ref: [6])
\item hydropathicity \code{"hphob.Eisenberg"} (Ref: [7])
\item hydropathicity \code{"hphob.HoppWoods"} (Ref: [8])
}}
}
\section{References}{

[1] Han S, Liang Y, Ma Q, \emph{et al}.
ncProR: an integrated R package for effective ncRNA-protein interaction prediction.
(\emph{Submitted})

[2] Morozova N, Allers J, Myers J, \emph{et al}.
Protein-RNA interactions: exploring binding patterns with a three-dimensional superposition analysis of high resolution structures.
Bioinformatics 2006; 22:2746-52

[3] Grantham R.
Amino acid difference formula to help explain protein evolution.
Science 1974; 185:862-4

[4] Zimmerman JM, Eliezer N, Simha R.
The characterization of amino acid sequences in proteins by statistical methods.
J. Theor. Biol. 1968; 21:170-201

[5] Bull HB, Breese K.
Surface tension of amino acid solutions: a hydrophobicity scale of the amino acid residues.
Arch. Biochem. Biophys. 1974; 161:665-670

[6] Kyte J, Doolittle RF.
A simple method for displaying the hydropathic character of a protein.
J. Mol. Biol. 1982; 157:105-132

[7] Eisenberg D, Schwarz E, Komaromy M, \emph{et al}.
Analysis of membrane and surface protein sequences with the hydrophobic moment plot.
J. Mol. Biol. 1984; 179:125-42

[8] Hopp TP, Woods KR.
Prediction of protein antigenic determinants from amino acid sequences.
Proc. Natl. Acad. Sci. U. S. A. 1981; 78:3824-8

[9] Kawashima S, Kanehisa M. AAindex: amino acid index database. Nucleic Acids Res. 2000; 28:374

[10] Bellucci M, Agostini F, Masin M, \emph{et al}.
Predicting protein associations with long noncoding RNAs.
Nat. Methods 2011; 8:444-445

[11] Lu Q, Ren S, Lu M, \emph{et al}.
Computational prediction of associations between long non-coding RNAs and proteins.
BMC Genomics 2013; 14:651
}

\examples{
data(demoPositiveSeq)
seqsRNA <- demoPositiveSeq$RNA.positive
seqsPro <- demoPositiveSeq$Pro.positive

# Return a data frame:
physChemRNA <- computePhysChem(seqs = seqsRNA, seqType = "RNA",
                               Fourier.len = 10, as.list = FALSE)

# Return a list:
physChemPro <- computePhysChem(seqs = seqsPro, seqType = "Pro", Fourier.len = 8,
                               physchemPro = c("polarity.Grantham",
                                               "polarity.Zimmerman",
                                               "hphob.BullBreese",
                                               "hphob.KyteDoolittle",
                                               "hphob.Eisenberg",
                                               "hphob.HoppWoods"),
                               as.list = TRUE)

}
\seealso{
\code{\link{featurePhysChem}}
}
