% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhysicochemicalProperty.R
\name{computePhysChem_AAindex}
\alias{computePhysChem_AAindex}
\title{Computation of the AAindex-Based Physicochemical Features of Protein Sequences}
\usage{
computePhysChem_AAindex(
  seqPro,
  entry.index = NULL,
  Fourier.len = 10,
  parallel.cores = 2,
  cl = NULL
)
}
\arguments{
\item{seqPro}{protein sequences loaded by function \code{\link[seqinr]{read.fasta}} from \code{\link[seqinr]{seqinr-package}}.
Each sequence should be a vector of single characters.}

\item{entry.index}{The entry number of AAindex. For example the entry number of "Kyte & Doolittle Hydrophaty"
index is 151, and the entry number of "Hopp-Woods Hydrophilicity" is 115. Thus,
the corresponding \code{entry.index} is \code{c(151, 115)} for these two physicochemical indices. See examples and \code{\link[seqinr]{aaindex}} of
"seqinr" package.}

\item{Fourier.len}{positive integer specifying the Fourier series length that will be used as features.
The \code{Fourier.len} should be >= the length of the input sequence. Default: \code{10}.}

\item{parallel.cores}{an integer that indicates the number of cores for parallel computation. Default: \code{2}.
Set \code{parallel.cores = -1} to run with all the cores. \code{parallel.cores} should be == -1 or >= 1.}

\item{cl}{parallel cores to be passed to this function.}
}
\value{
This function returns a data frame.
}
\description{
The function \code{computePhysChem_AAindex} computes the physicochemical features of
protein sequences with the help of AAindex provided by \code{\link[seqinr]{aaindex}} from \code{\link[seqinr]{seqinr-package}}.
}
\section{References}{

[1] Kawashima S, Kanehisa M.
AAindex: amino acid index database.
Nucleic Acids Res. 2000; 28:374

[2] Charif D, Lobry JR.
SeqinR 1.0-2: a contributed package to the R project for statistical computing devoted to biological sequences retrieval and analysis.
In: Structural approaches to sequence evolution: Molecules, networks, populations. 2007; 207-232
}

\examples{
data(demoPositiveSeq)
seqs <- demoPositiveSeq$Pro.positive
data(aaindex, package = "seqinr")

# Check the aaindex list provided by package "seqinr":

?seqinr::aaindex

# Supose that we need "Kyte & Doolittle Hydrophaty" index,
# and we can find the entries with Kyte as author:

which(sapply(aaindex, function(x) length(grep("Kyte", x$A)) != 0))

# This returns entry number 151.
# And 151 is the "entry.index" used by this function.
# Users can also obtain the entry number by retrieving information in other
# fields such as "x$D" (Data description) or "x$T" (Title of the article).
# See documentation of "seqinr::aaindex" for detailed information.

feature_aaindex_1 <- computePhysChem_AAindex(seqPro = seqs, entry.index = 151,
                                             Fourier.len = 10, parallel.cores = 2)

# If you need to compute features with more than aaindex data:

feature_aaindex_2 <- computePhysChem_AAindex(seqPro = seqs, entry.index = c(151, 68),
                                             Fourier.len = 10, parallel.cores = 2)

}
\seealso{
\code{\link{computePhysChem}}, \code{\link[seqinr]{aaindex}}
}
