% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{demoIDX}
\alias{demoIDX}
\title{A demo of sequence indices}
\format{
A data frame containing the names of RNA and protein sequences.
}
\usage{
data(demoIDX)
}
\description{
This data is an example of the input of \code{\link{formatSeq}}.
}
