% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{demoNegativeSeq}
\alias{demoNegativeSeq}
\title{A demo of negative pairs}
\usage{
data(demoNegativeSeq)
}
\description{
This file contains 20 negative RNA-protein interaction pairs. Required by some examples.
}
