% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{demoPositiveSeq}
\alias{demoPositiveSeq}
\title{A demo of positive pairs}
\usage{
data(demoPositiveSeq)
}
\description{
This file contains 20 positive RNA-protein interaction pairs. Required by some examples.
}
