% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelling.R
\name{evaluatePrediction}
\alias{evaluatePrediction}
\title{Evaluate Predicted Result}
\usage{
evaluatePrediction(reference, prediction, positive.class = NULL)
}
\arguments{
\item{reference}{a factor/character string of classes to be used as the true results.}

\item{prediction}{a factor/character string of predicted classes.}

\item{positive.class}{\code{NULL} or string. Which class is the positive class? Should be one
of the classes in label column. The first class in label column will be selected
as the positive class if leave \code{positive.class = NULL}.}
}
\value{
A dataframe that reports TP, TN, FP, FN, Sensitivity, Specificity, Accuracy,
F-Measure (F1-Score), MCC (Matthews Correlation Coefficient) and Cohen's Kappa.
}
\description{
This function can evaluate prediction based on reference labels and predicted results.
}
\details{
\code{reference} and \code{prediction} should have exactly the same classes.
More information please refer to \code{\link[caret]{confusionMatrix}}.
}
\section{References}{

Kuhn M.
Building predictive models in R using the caret package.
Journal of statistical software. 2008; 28(5):1-26.
}

\examples{

# Following codes only show how to use this function
# and cannot reflect the genuine performance of tools or classifiers.

data(demoPositiveSeq)
seqRNA <- demoPositiveSeq$RNA.positive
seqPro <- demoPositiveSeq$Pro.positive

# Predicting ncRNA-protein pairs using RPISeq (web-based):

Res_RPISeq <- run_RPISeq(seqRNA = seqRNA, seqPro = seqPro,
                         parallel.cores = 2) # Network is required.

# Evaluating the result:

perf_RPISeq <- evaluatePrediction(reference = rep("Non.Interact", 20),
                                  prediction = Res_RPISeq$RPISeq_Web_RF_pred,
                                  positive.class = "Non.Interact")

}
\seealso{
\code{\link[caret]{confusionMatrix}}
}
