% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Motifs.R
\name{featureMotifs}
\alias{featureMotifs}
\title{Extraction of the Motif Features of RNA and Protein Sequences}
\usage{
featureMotifs(
  seqRNA,
  seqPro,
  label = NULL,
  featureMode = c("concatenate", "combine"),
  newMotif.RNA = NULL,
  newMotif.Pro = NULL,
  newMotifOnly.RNA = FALSE,
  newMotifOnly.Pro = FALSE,
  parallel.cores = 2,
  cl = NULL,
  ...
)
}
\arguments{
\item{seqRNA}{RNA sequences loaded by function \code{\link[seqinr]{read.fasta}} from \code{\link[seqinr]{seqinr-package}}. Or a list of RNA sequences.
RNA sequences will be converted into lower case letters.}

\item{seqPro}{protein sequences loaded by function \code{\link[seqinr]{read.fasta}} from \code{\link[seqinr]{seqinr-package}}. Or a list of protein sequences.
Protein sequences will be converted into upper case letters.}

\item{label}{optional. A string or a vector of strings or \code{NULL}.
Indicates the class of the samples such as
"Interact", "Non.Interact". Default: \code{NULL}.}

\item{featureMode}{a string that can be \code{"concatenate"} or \code{"combine"}.
If \code{"concatenate"}, the motif features of RNA and proteins will be simply concatenated.
If \code{"combine"}, the returned dataset will be formed by combining the motif features of RNA and proteins.
See details below. Default: \code{"concatenate"}.}

\item{newMotif.RNA}{a list specifying the motifs that are counted in RNA sequences. Default: \code{NULL}.
For example, \code{newMotif = list(hnRNPA1 = c("UAGGGU", "UAGGGA"), SF1 = "UACUAAC")}.
Can be used with parameter \code{motifRNA} (see parameter \code{...}) to count motifs in RNA sequences.}

\item{newMotif.Pro}{a list specifying the motifs that are counted in protein sequences. Default: \code{NULL}.
For example, \code{newMotif = list(YGG = "YGG", E = "E")}.
Can be used with parameter \code{motifPro} (see parameter \code{...}) to count motifs in protein sequences.}

\item{newMotifOnly.RNA}{logical. If \code{TRUE}, only the new motifs defined in \code{newMotif.RNA} will be counted.
Default: \code{FALSE}.}

\item{newMotifOnly.Pro}{logical. If \code{TRUE}, only the new motifs defined in \code{newMotif.Pro} will be counted.
Default: \code{FALSE}.}

\item{parallel.cores}{an integer that indicates the number of cores for parallel computation. Default: \code{2}.
Set \code{parallel.cores = -1} to run with all the cores. \code{parallel.cores} should be == -1 or >= 1.}

\item{cl}{parallel cores to be passed to this function.}

\item{...}{argument \code{motifRNA} and \code{motifPro} to be passed to \code{\link{computeMotifs}}.
Used to compute the default motifs. See examples below.}
}
\value{
This function returns a data frame. Row names are the sequences names, and column names are the motif names.
The names of RNA and protein sequences are separated with ".",
i.e. row names format: "\emph{RNASequenceName.proteinSequenceName}" (e.g. "YDL227C.YOR198C").
If \code{featureMode = "combine"}, the motif names of RNA and protein sequences are also separated with ".",
i.e. column names format: "motif_\emph{RNAMotifName}.motif_\emph{proteinMotifName}" (e.g. "motif_PUM.motif_EE").
}
\description{
Basically a wrapper for \code{\link{computeMotifs}} function.
This function can count the motifs of RNA and protein sequences at the same time
and format the results as the dataset that can be used to build classifier.
}
\details{
If \code{featureMode = "concatenate"}, \emph{m} RNA motif features will be simply
concatenated with \emph{n} protein motif features, and the final result has \emph{m} + \emph{n} features.
If \code{featureMode = "combine"}, \emph{m} RNA motif features will be
combined with \emph{n} protein motif features, resulting in \emph{m} * \emph{n} possible combinations.

\code{...} can be used to pass the default motif patterns of RNA and protein sequences.
See arguments \code{motifRNA} and \code{motifPro} in \code{\link{computeMotifs}}.
}
\section{References}{

[1] Han S, \emph{et al}.
LION: an integrated R package for effective ncRNA-protein interaction prediction.
(\emph{Submitted})

[2] Akbaripour-Elahabad M, Zahiri J, Rafeh R, \emph{et al}.
rpiCOOL: A tool for In Silico RNA-protein interaction detection using random forest.
J. Theor. Biol. 2016; 402:1-8

[3] Pancaldi V, Bahler J.
In silico characterization and prediction of global protein-mRNA interactions in yeast.
Nucleic Acids Res. 2011; 39:5826-36

[4] Castello A, Fischer B, Eichelbaum K, \emph{et al}.
Insights into RNA Biology from an Atlas of Mammalian mRNA-Binding Proteins.
Cell 2012; 149:1393-1406

[5] Ray D, Kazan H, Cook KB, \emph{et al}.
A compendium of RNA-binding motifs for decoding gene regulation.
Nature 2013; 499:172-177

[6] Jiang P, Singh M, Coller HA.
Computational assessment of the cooperativity between RNA binding proteins and MicroRNAs in Transcript Decay.
PLoS Comput. Biol. 2013; 9:e1003075
}

\examples{
data(demoPositiveSeq)
seqsRNA <- demoPositiveSeq$RNA.positive
seqsPro <- demoPositiveSeq$Pro.positive

dataset1 <- featureMotifs(seqRNA = seqsRNA, seqPro = seqsPro, featureMode = "conc",
                          newMotif.RNA = list(motif1 = c("cc", "cu")),
                          newMotif.Pro = list(motif2 = "KK"),
                          motifRNA = c("Fusip1", "AU", "UG"),
                          motifPro = c("E", "K", "HR_RH"))

dataset2 <- featureMotifs(seqRNA = seqsRNA, seqPro = seqsPro, featureMode = "comb",
                          newMotif.RNA = list(motif1 = c("cc", "cu")),
                          newMotif.Pro = list(motif2 = c("R", "H")),
                          newMotifOnly.RNA = TRUE, newMotifOnly.Pro = FALSE)

}
\seealso{
\code{\link{computeMotifs}}
}
