% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Structure.R
\name{featureStructure}
\alias{featureStructure}
\title{Extraction of the Secondary Structural Features of RNA and Protein Sequences}
\usage{
featureStructure(
  seqRNA,
  seqPro,
  label = NULL,
  parallel.cores = 2,
  cl = NULL,
  ...
)
}
\arguments{
\item{seqRNA}{RNA sequences loaded by function \code{\link[seqinr]{read.fasta}} from \code{\link[seqinr]{seqinr-package}}. Or a list of RNA sequences.
RNA sequences will be converted into lower case letters.
Each sequence should be a vector of single characters.}

\item{seqPro}{protein sequences loaded by function \code{\link[seqinr]{read.fasta}} from \code{\link[seqinr]{seqinr-package}}. Or a list of protein sequences.
Protein sequences will be converted into upper case letters.
Each sequence should be a vector of single characters.}

\item{label}{optional. A string or a vector of strings or \code{NULL}.
Indicates the class of the samples such as
"Interact", "Non.Interact". Default: \code{NULL}.}

\item{parallel.cores}{an integer that indicates the number of cores for parallel computation.
Default: \code{2}. Set \code{parallel.cores = -1} to run with all the cores. \code{parallel.cores} should be == -1 or >= 1.}

\item{cl}{parallel cores to be passed to this function.}

\item{...}{arguments (\code{structureRNA.num}, \code{structurePro}, \code{Fourier.len},
\code{workDir.Pro}, \code{path.RNAsubopt}, \code{path.Predator} and \code{path.stride}
passed to function \code{\link{computeStructure}}. See example below.}
}
\value{
This function returns a data frame.
}
\description{
Basically a wrapper for \code{\link{computeStructure}} function.
This function can extract secondary structure features of RNA and protein sequences at the same time
and format the results as the dataset that can be used to build classifier.
ViennaRNA package and Predator is required.
}
\details{
see \code{\link{computeStructure}}.
}
\section{References}{

[1] Han S, Liang Y, Ma Q, \emph{et al}.
ncProR: an integrated R package for effective ncRNA-protein interaction prediction.
(\emph{Submitted})

[2] Chou PY, Fasman GD.
Prediction of the secondary structure of proteins from their amino acid sequence.
Adv. Enzymol. Relat. Areas Mol. Biol. 1978; 47:45-148

[3] Deleage G, Roux B.
An algorithm for protein secondary structure prediction based on class prediction.
Protein Eng. Des. Sel. 1987; 1:289-294

[4] Levitt M.
Conformational preferences of amino acids in globular proteins.
Biochemistry 1978; 17:4277-85

[5] Frishman D, Argos P.
Incorporation of non-local interactions in protein secondary structure prediction from the amino acid sequence.
Protein Eng. 1996; 9:133-42

[6] Lorenz R, Bernhart SH, Honer zu Siederdissen C, \emph{et al}.
ViennaRNA Package 2.0.
Algorithms Mol. Biol. 2011; 6:26

[7] Lu Q, Ren S, Lu M, \emph{et al}.
Computational prediction of associations between long non-coding RNAs and proteins.
BMC Genomics 2013; 14:651
}

\examples{
\dontrun{
data(demoNegativeSeq)
seqsRNA <- demoNegativeSeq$RNA.negative
seqsPro <- demoNegativeSeq$Pro.negative

# Use your own paths of the program Predator and file "stride.dat". For example:

path.Predator <- "/mnt/external_drive_1/hansy/predator/predator"
path.stride <- "/mnt/external_drive_1/hansy/predator/stride.dat"

# Pass "structureRNA.num", "structurePro", "Fourier.len", "workDir.Pro",
# "path.RNAsubopt", "path.Predator" and "path.stride" using "..." argument:

dataset <- featureStructure(seqRNA = seqsRNA, seqPro = seqsPro, label = "Non.Interact",
                            parallel.cores = 2, structureRNA.num = 6,
                            structurePro = c("ChouFasman", "DeleageRoux", "Levitt"),
                            Fourier.len = 10, workDir.Pro = "tmpDir",
                            path.RNAsubopt = "RNAsubopt", path.Predator = path.Predator,
                            path.stride = path.stride)
}
}
\seealso{
\code{\link{runRNAsubopt}}, \code{\link{runPredator}}, \code{\link{computeStructure}}
}
