% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelling.R
\name{formatSeq}
\alias{formatSeq}
\title{Format RNA/Protein Sequences According to the Index}
\usage{
formatSeq(idx, seqs)
}
\arguments{
\item{idx}{specifiying the sequence indices.}

\item{seqs}{sequences loaded by function \code{\link[seqinr]{read.fasta}} of package
"seqinr" (\code{\link[seqinr]{seqinr-package}}). Or a list of sequences.}
}
\value{
This function returns a list.
}
\description{
This function generates a list of sequences according to the specified indices.
The sequence list can be used as input for feature extraction or prediction.
}
\details{
This function is useful for generating the dataset with the specified indices and a sequence list.
For example, the positive pairs of RNA-protein interactions have been provided, but the sequences
are randomly listed in one file. This function can generate the sequence list according to the indices.
See examples for more details.
}
\examples{
data(demoIDX)
data(demoPositiveSeq)

new_RNA <- formatSeq(demoIDX$RNA_index, demoPositiveSeq$RNA.positive)
new_Pro <- formatSeq(demoIDX$Pro_index, demoPositiveSeq$Pro.positive)

names(demoPositiveSeq$Pro.positive)
names(demoPositiveSeq$RNA.positive)

names(new_RNA)
names(new_Pro)

# new_RNA and new_Pro can be further used to extract features.


}
