% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mod_LION}
\alias{mod_LION}
\title{Retrained random forest classifier of LION method}
\usage{
data(mod_LION)
}
\description{
This file is required by function \code{\link{run_LION}} to predict ncRNA-protein interaction.
Users don't need to load this file manually.
}
