% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mod_LncADeep}
\alias{mod_LncADeep}
\title{Retrained classifier of LncADeep method}
\usage{
data(mod_LncADeep)
}
\description{
This file is required by function \code{\link{run_LncADeep}} to predict ncRNA-protein interaction.
Users don't need to load this file manually.
}
