% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mod_lncPro}
\alias{mod_lncPro}
\title{Retrained random forest classifier of lncPro method}
\usage{
data(mod_lncPro)
}
\description{
This file is required by function \code{\link{run_lncPro}} to predict ncRNA-protein interaction.
Users don't need to load this file manually.
}
