% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mod_rpiCOOL}
\alias{mod_rpiCOOL}
\title{Retrained random forest classifier of rpiCOOL method}
\usage{
data(mod_rpiCOOL)
}
\description{
This file is required by function \code{\link{run_rpiCOOL}} to predict ncRNA-protein interaction.
Users don't need to load this file manually.
}
