% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{mod_RPISeq}
\alias{mod_RPISeq}
\title{Retrained random forest classifier of RPISeq method}
\usage{
data(mod_RPISeq)
}
\description{
This file is required by function \code{\link{run_RPISeq}} to predict ncRNA-protein interaction.
Users don't need to load this file manually.
}
