% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modelling.R
\name{normalizeData}
\alias{normalizeData}
\title{Normalize Data}
\usage{
normalizeData(dataset, direction = c("row", "column"),
  returnNorm = TRUE, ignoreColumn = NULL)
}
\arguments{
\item{dataset}{input dataset. As a dataframe.}

\item{direction}{\code{"row"} or {"column"} to indicate the normalization direction. (see details)}

\item{returnNorm}{logical. Used for \code{direction = "column"}. The function will return nomalized
data that can be used to process new dataset (test sets, for example).
Ignored when \code{direction = "row"}.}

\item{ignoreColumn}{numeric or \code{NULL}. Input the column number of the dataset
to indicate which column(s) will not be normalized. Default: \code{NULL}.}
}
\value{
If \code{direction = "column"} and \code{returnNorm = TRUE},
the function returns a list containing normalized dataset and normalization data.
Otherwise, only return the processed dataset.
}
\description{
This function is used to normalize dataset.
}
\details{
The function provides two normalization strategies: by row or by column.
If by row, the dataset will be processed with equation (see reference):
di = (fi - min\{f1, f2, ...\}) / max\{f1, f2, ...\}.
f1, f2, ..., fi are the original values of each row.

If by column, the dataset will be processed with:
di = (fi - min\{f1, f2, ...\}) / (max\{f1, f2, ...\} - min\{f1, f2, ...\}).
}
\section{References}{

Shen J, Zhang J, Luo X, \emph{et al}.
Predicting protein-protein interactions based only on sequences information.
Proc. Natl. Acad. Sci. U. S. A. 2007; 104:4337-41
}

\examples{

data(demoPositiveSeq)
seqRNA <- demoPositiveSeq$RNA.positive
seqPro <- demoPositiveSeq$Pro.positive

dataset <- featureFreq(seqRNA = seqRNA, seqPro = seqPro, label = "Interact",
                       featureMode = "conc", computePro = "DeNovo", k.Pro = 3,
                       k.RNA = 2, normalize = "none", parallel.cores = 2)

processed_1 <- normalizeData(dataset, direction = "row", ignoreColumn = 1)
processed_2 <- normalizeData(dataset[,-1], direction = "column",
                             returnNorm = TRUE, ignoreColumn = NULL)

}
