% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{run_confidentPrediction}
\alias{run_confidentPrediction}
\title{Confident Prediction of RNA-Protein Interaction Using Multiple Methods Simultaneously}
\usage{
run_confidentPrediction(
  seqRNA,
  seqPro,
  label = NULL,
  methods = c("RPISeq_web", "RPISeq_retrain", "lncPro_original", "lncPro_retrain",
    "rpiCOOL_retrain", "LncADeep_retrain", "LION"),
  RPISeq.mod = NULL,
  lncPro.mod = NULL,
  rpiCOOL.mod = NULL,
  LncADeep.mod = NULL,
  LION.mod = NULL,
  parallel.cores = 2,
  cl = NULL
)
}
\arguments{
\item{seqRNA}{RNA sequences loaded by function \code{\link[seqinr]{read.fasta}} from \code{\link[seqinr]{seqinr-package}}. Or a list of RNA/protein sequences.
RNA sequences will be converted into lower case letters.}

\item{seqPro}{protein sequences loaded by function \code{\link[seqinr]{read.fasta}} from \code{\link[seqinr]{seqinr-package}}. Or a list of protein sequences.
Protein sequences will be converted into upper case letters.
Each sequence should be a vector of single characters.}

\item{label}{optional. A string or a vector of strings or \code{NULL}. Used to give labels or notes to the output result.
Default: \code{NULL}.}

\item{methods}{strings. Indicate the method(s) to be used for prediction.
Can be: \code{"RPISeq_web"}, \code{"RPISeq_retrain"}, \code{"lncPro_original"},
\code{"lncPro_retrain"}, \code{"rpiCOOL_retrain"}, \code{"LncADeep_retrain"} and \code{"LION"}.}

\item{RPISeq.mod, lncPro.mod, rpiCOOL.mod, LncADeep.mod, LION.mod}{use default retrained model (if \code{NULL}) or assign a new retrained model?
New retrained model can be generated with \code{\link{run_RPISeq}}, \code{\link{run_lncPro}}, \code{\link{run_rpiCOOL}}, \code{\link{run_LncADeep}} and \code{\link{LION}}.}

\item{parallel.cores}{an integer that indicates the number of cores for parallel computation.
Default: \code{2}. Set \code{parallel.cores = -1} to run with all the cores. \code{parallel.cores} should be == -1 or >= 1.}

\item{cl}{parallel cores to be passed to this function.}
}
\value{
A list containing the predicted results.
}
\description{
This function can predict lncRNA/RNA-protein interactions using all supported methods, which is useful to have a high-confident prediction.
}
\section{References}{

Han S, Liang Y, Ma Q, \emph{et al}.
LION: an integrated R package for effective ncRNA-protein interaction prediction.
(\emph{Submitted})
}

\examples{

#' # Following codes only show how to use this function
# and cannot reflect the genuine performance of tools or classifiers.

data(demoPositiveSeq)
seqRNA <- demoPositiveSeq$RNA.positive
seqPro <- demoPositiveSeq$Pro.positive

# Using methods RPISeq (retrained model) and rpiCOOL (retrained model):

Res_confidence <- run_confidentPrediction(seqRNA = seqRNA, seqPro = seqPro,
                                          methods = c("RPISeq_retrain", "rpiCOOL_retrain", "LION"),
                                          label = "Interact", # label is optional
                                          parallel.cores = 2)
# Convert to data frame:

Res_confidence_df <- do.call("cbind", Res_confidence)
Res_confidence_df <- Res_confidence_df[!duplicated(names(Res_confidence_df))]

}
