% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{run_lncPro}
\alias{run_lncPro}
\title{Predict RNA-Protein Interaction Using lncPro Method}
\usage{
run_lncPro(seqRNA, seqPro, mode = c("original", "retrained"),
  label = NULL, path.RNAsubopt = "RNAsubopt",
  path.Predator = "Predator/predator",
  path.stride = "Predator/stride.dat", workDir.Pro = getwd(),
  parallel.cores = 2)
}
\arguments{
\item{seqRNA}{RNA sequences loaded by function \code{\link[seqinr]{read.fasta}} of package
"seqinr" (\code{\link[seqinr]{seqinr-package}}). Or a list of RNA/protein sequences.
RNA sequences will be converted into lower case letters.}

\item{seqPro}{protein sequences loaded by function \code{\link[seqinr]{read.fasta}} of package
"seqinr" (\code{\link[seqinr]{seqinr-package}}). Or a list of protein sequences.
Protein sequences will be converted into upper case letters.
Each sequence should be a vector of single characters.}

\item{mode}{set \code{"original"} to use original lncPro algorithm, and \code{"retrained"} to call retrained model.
The retrained model is constructed with the same features as the original version, but random classifier is employed to
build classifier.}

\item{label}{string or \code{NULL}. Used to give label or just a note to the output result. Default: \code{NULL}.}

\item{path.RNAsubopt}{string specifying the location of "RNAsubopt" program.}

\item{path.Predator}{string specifying the location of "Predator" program.}

\item{path.stride}{string specifying the location of file "stride.dat" required by program Predator.}

\item{workDir.Pro}{string specifying the directory for temporary files.
The temp files will be deleted automatically when
the calculation is completed. If the directory does not exist, it will be created automatically.}

\item{parallel.cores}{an integer that indicates the number of cores for parallel computation.
Default: \code{2}. Set \code{parallel.cores = -1} to run with all the cores.}
}
\value{
This function returns a data frame that contains the predicted results.
}
\description{
This function can predict lncRNA/RNA-protein interactions using lncPro method.
Programs "RNAsubopt" from software "ViennaRNA Package" and "Predator" is required. Please also note that
"Predator" is only available on UNIX/Linux and 32-bit Windows OS.
}
\details{
The method is proposed by lncPro. This function, \code{runlncPro}, has
improved and fixed the original code.

\code{runlncPro} depends on the program "RNAsubopt" of software "ViennaRNA".
(\url{http://www.tbi.univie.ac.at/RNA/index.html})

Parameter \code{path.RNAsubopt} can be simply defined as \code{"RNAsubopt"} as
default when the OS is UNIX/Linux. However, for some OS, such as Windows, users may
need to specify the \code{path.RNAsubopt} if the path of "RNAsubopt" haven't been
added in environment variables (e.g. \code{path.RNAsubopt = '"C:/Program Files/ViennaRNA/RNAsubopt.exe"'}).

Program "Predator" is only available on UNIX/Linux and 32-bit Windows OS.
}
\section{References}{

Lu Q, Ren S, Lu M, \emph{et al}.
Computational prediction of associations between long non-coding RNAs and proteins.
BMC Genomics 2013; 14:651
}

\examples{
\dontrun{
data(demoPositiveSeq)
seqRNA <- demoPositiveSeq$RNA.positive
seqPro <- demoPositiveSeq$Pro.positive

Res_lncPro <- run_lncPro(seqRNA = seqRNA, seqPro = seqPro,
                         path.RNAsubopt = "RNAsubopt",
                         path.Predator = "~/predator/predator",
                         path.stride = "~/predator/stride.dat",
                         workDir.Pro = "tmp", parallel.cores = 10)
}

}
