% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{run_ncProR}
\alias{run_ncProR}
\title{Predict RNA-Protein Interaction Using ncProR Method}
\usage{
run_ncProR(seqRNA, seqPro, label = NULL, parallel.cores = 2)
}
\arguments{
\item{seqRNA}{RNA sequences loaded by function \code{\link[seqinr]{read.fasta}} of package
"seqinr" (\code{\link[seqinr]{seqinr-package}}). Or a list of RNA/protein sequences.
RNA sequences will be converted into lower case letters.}

\item{seqPro}{protein sequences loaded by function \code{\link[seqinr]{read.fasta}} of package
"seqinr" (\code{\link[seqinr]{seqinr-package}}). Or a list of protein sequences.
Protein sequences will be converted into upper case letters.
Each sequence should be a vector of single characters.}

\item{label}{string or \code{NULL}. Used to give label or just a note to the output result. Default: \code{NULL}.}

\item{parallel.cores}{an integer that indicates the number of cores for parallel computation.
Default: \code{2}. Set \code{parallel.cores = -1} to run with all the cores.}
}
\value{
This function returns a data frame that contains the predicted results.
}
\description{
This function can predict lncRNA/RNA-protein interactions using ncProR method.
}
\section{References}{

Han S, Liang Y, Li Y, \emph{et al}.
ncProR: an integrated R package for effective ncRNA-protein interaction prediction.
(\emph{Submitted})
}

\examples{

data(demoPositiveSeq)
seqRNA <- demoPositiveSeq$RNA.positive
seqPro <- demoPositiveSeq$Pro.positive

Res_ncProR <- run_ncProR(seqRNA = seqRNA, seqPro = seqPro, parallel.cores = 2)


}
