% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{run_rpiCOOL}
\alias{run_rpiCOOL}
\title{Predict RNA-Protein Interaction Using rpiCOOL Method}
\usage{
run_rpiCOOL(seqRNA, seqPro, label = NULL, parallel.cores = 2)
}
\arguments{
\item{seqRNA}{RNA sequences loaded by function \code{\link[seqinr]{read.fasta}} of package
"seqinr" (\code{\link[seqinr]{seqinr-package}}). Or a list of RNA/protein sequences.
RNA sequences will be converted into lower case letters.}

\item{seqPro}{protein sequences loaded by function \code{\link[seqinr]{read.fasta}} of package
"seqinr" (\code{\link[seqinr]{seqinr-package}}). Or a list of protein sequences.
Protein sequences will be converted into upper case letters.
Each sequence should be a vector of single characters.}

\item{label}{string or \code{NULL}. Used to give label or just a note to the output result. Default: \code{NULL}.}

\item{parallel.cores}{an integer that indicates the number of cores for parallel computation.
Default: \code{2}. Set \code{parallel.cores = -1} to run with all the cores.}
}
\value{
This function returns a data frame that contains the predicted results.
}
\description{
This function can predict lncRNA/RNA-protein interactions using retrained model of rpiCOOL method.
The codes of this function slightly differ from the rpiCOOL's script.
}
\section{References}{

Akbaripour-Elahabad M, Zahiri J, Rafeh R, \emph{et al}.
rpiCOOL: A tool for In Silico RNA-protein interaction detection using random forest.
J. Theor. Biol. 2016; 402:1-8
}

\examples{

data(demoPositiveSeq)
seqRNA <- demoPositiveSeq$RNA.positive
seqPro <- demoPositiveSeq$Pro.positive

Res_rpiCOOL <- run_rpiCOOL(seqRNA = seqRNA, seqPro = seqPro,
                           label = "rpiCOOL_res", parallel.cores = 2)


}
