% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{run_RPISeq}
\alias{run_RPISeq}
\title{Predict RNA-Protein Interaction Using RPISeq Method}
\usage{
run_RPISeq(seqRNA, seqPro, mode = c("web", "retrained"), label = NULL,
  parallel.cores = 2)
}
\arguments{
\item{seqRNA}{RNA sequences loaded by function \code{\link[seqinr]{read.fasta}} of package
"seqinr" (\code{\link[seqinr]{seqinr-package}}). Or a list of RNA/protein sequences.
RNA sequences will be converted into lower case letters.}

\item{seqPro}{protein sequences loaded by function \code{\link[seqinr]{read.fasta}} of package
"seqinr" (\code{\link[seqinr]{seqinr-package}}). Or a list of protein sequences.
Protein sequences will be converted into upper case letters.
Each sequence should be a vector of single characters.}

\item{mode}{set \code{"web"} to employ web-based original version.
Use \code{"retrained"} to call retrained model.}

\item{label}{string or \code{NULL}. Used to give label or just a note to the output result. Default: \code{NULL}.}

\item{parallel.cores}{an integer that indicates the number of cores for parallel computation.
Default: \code{2}. Set \code{parallel.cores = -1} to run with all the cores.}
}
\value{
This function returns a data frame that contains the predicted results.
}
\description{
This function can predict lncRNA/RNA-protein interactions using RPISeq method.
Both the web-based original version and retrained model are available. Network is required to use
the original version.
}
\section{References}{

Muppirala UK, Honavar VG, Dobbs D.
Predicting RNA-protein interactions using only sequence information.
BMC Bioinformatics 2011; 12:489
}

\examples{

data(demoPositiveSeq)
seqRNA <- demoPositiveSeq$RNA.positive
seqPro <- demoPositiveSeq$Pro.positive

Res_RPISeq_1 <- run_RPISeq(seqRNA = seqRNA, seqPro = seqPro, mode = "web",
                           parallel.cores = 2)

Res_RPISeq_2 <- run_RPISeq(seqRNA = seqRNA, seqPro = seqPro, mode = "retrained",
                           label = "Interact", parallel.cores = 2)


}
