% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Structure.R
\name{runPredator}
\alias{runPredator}
\title{Run Predator in R}
\usage{
runPredator(seqs, path.Predator, path.stride, workDir = getwd(),
  verbose = FALSE, parallel.cores = 2)
}
\arguments{
\item{seqs}{RNA sequences loaded by function \code{\link[seqinr]{read.fasta}} of package
"seqinr" (\code{\link[seqinr]{seqinr-package}}). Or a list of RNA/protein sequences.
RNA sequences will be converted into lower case letters.
Each sequence should be a vector of single characters. (Non-AA letters will be ignored.)}

\item{path.Predator}{string specifying the location of Predator program.}

\item{path.stride}{string specifying the location of file "stride.dat" required by program Predator.}

\item{workDir}{string specifying the directory for temporary files.
The temp files will be deleted automatically when
the calculation is completed. The directory does not exist will be created automatically.}

\item{verbose}{logical. Should the relevant information be printed during the calculation? (Only available on UNIX/Linux.)}

\item{parallel.cores}{an integer that indicates the number of cores for parallel computation. Default: \code{2}.
Set \code{parallel.cores = -1} to run with all the cores.}
}
\value{
This function returns a data frame of the raw outputs of Predator.
}
\description{
This function can run and capture the result of Predator in R
by invoking the OS command.
The results can be formatted and used as the features of tool "lncPro".
Predator is required (NOTE: Predator does not support 64-bit MS Windows OS. Linux OS is recommended).
}
\details{
This function depends on the program Predator.
Program Predator is only available on UNIX/Linux and 32-bit Windows OS.

This function can print the related information when the OS is UNIX/Linux,
such as:

\code{"25 of 100, length: 50 aa"},

which means around 100 sequences are assigned to this node, and the program is
computing the 25th sequence. The length of this sequence is 50 aa.
}
\section{References}{

Frishman D, Argos P.
Incorporation of non-local interactions in protein secondary structure prediction from the amino acid sequence.
Protein Eng. 1996; 9:133-42
}

\examples{
\dontrun{
data(demoPositiveSeq)
seqsPro <- demoPositiveSeq$Pro.positive

path.predator = "/mnt/external_drive_1/hansy/R_tmp/LncProInter/Predator/predator"
path.stride = "/mnt/external_drive_1/hansy/R_tmp/LncProInter/Predator/stride.dat"

Predator.res <- runPredator(seqs = Pro.positive, path.Predator = path.Predator,
                            path.stride = path.stride, workDir = "tmp",
                            as.string = TRUE, parallel.cores = 4)
}
}
\seealso{
\code{\link{runRNAsubopt}}
}
