% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\name{run_ncProR}
\alias{run_ncProR}
\title{Predict RNA-Protein Interaction Using ncProR Method}
\usage{
run_ncProR(
  seqRNA,
  seqPro,
  mode = c("prediction", "retrain", "feature"),
  retrained.model = NULL,
  label = NULL,
  positive.class = NULL,
  folds.num = 10,
  ntree.range = c(200, 500, 1000, 1500, 2000),
  seed = 1,
  parallel.cores = 2,
  ...
)
}
\arguments{
\item{seqRNA}{RNA sequences loaded by function \code{\link[seqinr]{read.fasta}} from \code{\link[seqinr]{seqinr-package}}. Or a list of RNA/protein sequences.
RNA sequences will be converted into lower case letters.}

\item{seqPro}{protein sequences loaded by function \code{\link[seqinr]{read.fasta}} from \code{\link[seqinr]{seqinr-package}}. Or a list of protein sequences.
Protein sequences will be converted into upper case letters.
Each sequence should be a vector of single characters.}

\item{mode}{a string. Set \code{"prediction"} to predict ncRNA-protein pairs and return prediction results;
set \code{"retrain"} to build a new random forest model using the input data;
set \code{"feature"} to return a data frame contains the extracted features.
Users can use the extracted features generated by \code{mode = "feature"} to train classifiers
with other machine learning algorithms. Default: \code{"prediction"}.}

\item{retrained.model}{(only when \code{mode = "prediction"})
use the default model or a new retrained model to predict ncRNA-protein pairs?
If \code{NULL}, default machine learning model will be used. Or pass the model generated by this function
with parameter \code{"mode = retrain"}. Default: \code{NULL}. See examples below.}

\item{label}{a string or a vector of strings or \code{NULL}.
Optional when \code{mode = "prediction"} or \code{mode = "feature"}: used to give labels or notes to the output result.
Required when \code{mode = "retrain"}: must be a vector of strings that corresponds to input sequences.
Each string indicates the class of each input pair. Default: \code{NULL}.}

\item{positive.class}{(only when \code{mode = "retrain"}) \code{NULL} or a string used to indicate
which class is the positive class, Should be one
of the classes in \code{label} or leave \code{positive.class = NULL}.
In the latter case, the first class in \code{label} will be used
as the positive class. Default: \code{NULL}.}

\item{folds.num}{(only when \code{mode = "retrain"}) an integer indicates the number of folds for cross validation.
Default: \code{10} for 10-fold cross validation.}

\item{ntree.range}{(only when \code{mode = "retrain"}) used to indicate the range of \code{ntree}
when tuning the random forest classifier.
Default: \code{c(200, 500, 1000, 1500, 2000)}.}

\item{seed}{(only when \code{mode = "retrain"}) an integer indicates the random seed for data splitting.}

\item{parallel.cores}{an integer that indicates the number of cores for parallel computation.
Default: \code{2}. Set \code{parallel.cores = -1} to run with all the cores. \code{parallel.cores} should be == -1 or >= 1.}

\item{...}{(only when \code{mode = "retrain"}) other parameters passed to \code{\link[randomForest]{randomForest}} function.}
}
\value{
If \code{mode = "prediction"}, this function returns a data frame that contains the predicted results.
If \code{mode = "retrain"}, this function returns a random forest classifier.
If \code{mode = "prediction"}, this function returns a data frame that contains the extracted features.
}
\description{
This function can predict lncRNA/RNA-protein interactions using ncProR method. Model retraining and feature extraction are also supported.
}
\section{References}{

Han S, Liang Y, Ma Q, \emph{et al}.
ncProR: an integrated R package for effective ncRNA-protein interaction prediction.
(\emph{Submitted})
}

\examples{

# Following codes only show how to use this function
# and cannot reflect the genuine performance of tools or classifiers.

data(demoPositiveSeq)
seqRNA <- demoPositiveSeq$RNA.positive
seqPro <- demoPositiveSeq$Pro.positive

# Predicting ncRNA-protein pairs:

Res_ncProR_1 <- run_ncProR(seqRNA = seqRNA, seqPro = seqPro,
                           parallel.cores = 2) # using the default setting

# the above command is equivalent to:
Res_ncProR_2 <- run_ncProR(seqRNA = seqRNA, seqPro = seqPro, mode = "prediction",
                           retrained.model = NULL, label = NULL,
                           parallel.cores = 2)

# Train a new model:

# Argument "label" which indicates the class of each input pair is required here.
# "label" should correspond to the classes of "seqRNA" and "seqPro".
# "positive.class" should be one of the classes in argument "label" or can be set as "NULL".
# In the latter case, the first label in "label" will be used as the positive class.
# Parameters of random forest, such as "mtry", can be passed using "..." argument.

ncProR_model <- run_ncProR(seqRNA = seqRNA, seqPro = seqPro, mode = "retrain",
                           label = rep(c("Interact", "Non.Interact"), each = 10),
                           positive.class = NULL, folds.num = 5,
                           ntree.range = c(300, 500), seed = 1,
                           parallel.cores = 2, mtry = 20)

# Predicting using new built model by setting "retrained.model = ncProR_model":

Res_ncProR_2 <- run_ncProR(seqRNA = seqRNA, seqPro = seqPro, mode = "prediction",
                           retrained.model = ncProR_model,
                           label = rep(c("Interact", "Non.Interact"), each = 10),
                           parallel.cores = 2)

# Only extracting features:

ncProR_feature_df <- run_ncProR(seqRNA = seqRNA, seqPro = seqPro, mode = "feature",
                                label = "ncProR_feature", parallel.cores = 2)

# Extracted features can be used to build classifiers using other machine learning
# algorithms, which provides users with more flexibility.

}
